requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
requests>=2.31
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0